legacy/requirements.txt
requests
python-dotenv
loguru
beautifulsoup4
playwright
